requirements.txt
PyMySQL==1.0.2
python-dotenv==0.17.1
